Add files via upload
The purpose of the Online Mental Health Support Group website is to create a platform where individuals can connect and discuss their mental health issues in a supportive and understanding environment. This is a team project with five members, and I contribute as a Frontend Developer while also managing the database.